+++ Moda.Web/src/moda.web.reactclient/src/app/components/common/planning/planning-interval-objectives-grid.tsx
@@ -7,7 +7,7 @@ import ModaGrid from '../moda-grid'
 import { MenuProps, Progress, Space, Switch } from 'antd'
 import { ItemType } from 'antd/es/menu/hooks/useItems'
 import useAuth from '../../contexts/auth'
-import EditPlanningIntervalObjectiveForm from '@/src/app/planning/planning-intervals/[key]/edit-planning-interval-objective-form'
+import EditPlanningIntervalObjectiveForm from '@/src/app/planning/components/edit-planning-interval-objective-form'
 import { UseQueryResult } from 'react-query'
 import dayjs from 'dayjs'
 import CreateHealthCheckForm from '../health-check/create-health-check-form'
