+++ Moda.Web/src/moda.web.reactclient/src/app/planning/components/index.ts
@@ -1,3 +1,5 @@
 export { default as CreatePlanningIntervalForm } from './create-planning-interval-form'
 export { default as EditPlanningIntervalForm } from './edit-planning-interval-form'
 export { default as PlanningIntervalObjectivesTimeline } from './planning-interval-objectives-timeline'
+export { default as CreatePlanningIntervalObjectiveForm } from './create-planning-interval-objective-form'
+export { default as EditPlanningIntervalObjectiveForm } from './edit-planning-interval-objective-form'
