+++ Moda.Web/src/moda.web.reactclient/src/app/planning/planning-intervals/[key]/loading.tsx
@@ -6,7 +6,7 @@ import { Skeleton } from 'antd'
 export default function PlanningIntervalDetailsLoading() {
   return (
     <>
-      <PageTitle title="PI Details" />
+      <PageTitle title="Planning Interval" />
       <Skeleton active />
     </>
   )
