+++ Moda.Web/src/moda.web.reactclient/src/app/planning/planning-intervals/[key]/objectives/[objectiveKey]/page.tsx
@@ -6,7 +6,7 @@ import { useEffect, useMemo, useState } from 'react'
 import PlanningIntervalObjectiveDetails from './planning-interval-objective-details'
 import { useDocumentTitle } from '@/src/app/hooks/use-document-title'
 import useAuth from '@/src/app/components/contexts/auth'
-import EditPlanningIntervalObjectiveForm from '../../edit-planning-interval-objective-form'
+import EditPlanningIntervalObjectiveForm from '../../../../components/edit-planning-interval-objective-form'
 import { DownOutlined } from '@ant-design/icons'
 import { ItemType } from 'antd/es/menu/hooks/useItems'
 import DeletePlanningIntervalObjectiveForm from './delete-planning-interval-objective-form'
