+++ Moda.Web/src/moda.web.reactclient/src/app/planning/planning-intervals/[key]/plan-review/team-objectives-list-card.tsx
@@ -7,7 +7,7 @@ import ObjectiveListItem from './objective-list-item'
 import ModaEmpty from '@/src/app/components/common/moda-empty'
 import { useCallback, useMemo, useState } from 'react'
 import useAuth from '@/src/app/components/contexts/auth'
-import CreatePlanningIntervalObjectiveForm from '../create-planning-interval-objective-form'
+import CreatePlanningIntervalObjectiveForm from '../../../components/create-planning-interval-objective-form'
 import dayjs from 'dayjs'
 import { UseQueryResult } from 'react-query'
 import useTheme from '@/src/app/components/contexts/theme'
