+++ Moda.Web/src/moda.web.reactclient/src/app/planning/planning-intervals/[key]/risks/loading.tsx
@@ -3,10 +3,10 @@
 import PageTitle from '@/src/app/components/common/page-title'
 import { Skeleton } from 'antd'
 
-export default function PlanningIntervalRisksPageLoading() {
+export default function PlanningIntervalObjectivesPageLoading() {
   return (
     <>
-      <PageTitle title="PI Risks" />
+      <PageTitle title="PI Objectives" />
       <Skeleton active />
     </>
   )
